Create journal_eigenfactors
I went through the output txt files you put together for us before.
At the end is a pandas dataframe that shows unique journals, and the frequency with which they occur. The last frame is subsetted for journals that are published in more than 100 times. The 100 is arbitrary, but I think we should set some cutoff, for obvious reasons. 
I've added eigenfactor values for some of the journals, but not all yet. Let me know if you have any thoughts. Thanks.