Display Histograms of data frame
This is for the distribution of all the stats that have been collected
on the small text file. I was able to run this on my computer. I think
the only thing I need to figure out is how to make the range smaller
for some of the graphs. Also I disregarded any cell that had a string
in them, which is the None statement from the aggregations.